made it up to edges

+++ deeplynx.models/RequestDTOs/EdgeRequestDto.cs
@@ -6,12 +6,12 @@ namespace deeplynx.models;
 public class EdgeRequestDto
 {
     [Required]
-    public int OriginId { get; set; } 
+    public long OriginId { get; set; } 
     
     [Required]
-    public int DestinationId { get; set; }
+    public long DestinationId { get; set; }
     
-    public int? RelationshipId { get; set; }
+    public long? RelationshipId { get; set; }
     
     public string? RelationshipName { get; set; }
 }
